usemem_remap: use unsigned long for pagesize
There are several places for pagesize doing arithmetic operations and
due to its int type, the result could overflow and yielding strange
result like end is calculated with a smaller value than start, etc.
The subsequent file_remap_pages will then fail with -EINVAL.

I also changed the type of no_of_pages from int to unsigned long for
similar reasons(although it's not required for this fix).

https://jira01.devtools.intel.com/browse/OLT-1605
Reported-by: Philip Li <philip.li@intel.com>

+++ usemem_remap.c
@@ -18,11 +18,11 @@
 int main (int argc, char *argv[])
 {
 	int fd;
-	int pagesize = getpagesize();
+	unsigned long pagesize = getpagesize();
 	int i;
 
 	unsigned long size;
-	int no_of_pages;
+	unsigned long no_of_pages;
 
 	/* initialize to NULL */
 	unsigned char *start = NULL;
